+++ src/main/java/org/sopt/post/presentation/PostController.java
@@ -12,8 +12,8 @@ import static org.sopt.post.presentation.message.PostMessage.UPDATED_SUCCESS;
 import jakarta.validation.Valid;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
-import org.sopt.post.application.request.CreatePostServiceRequest;
-import org.sopt.post.application.request.UpdatePostServiceRequest;
+import org.sopt.post.application.dto.request.CreatePostServiceRequest;
+import org.sopt.post.application.dto.request.UpdatePostServiceRequest;
 import org.sopt.post.presentation.dto.request.CreatePostRequest;
 import org.sopt.post.presentation.dto.request.UpdatePostRequest;
 import org.sopt.post.presentation.dto.response.GetAllPostsResponse;
